locking down users mutation

+++ src/resolvers/Mutation.js
@@ -45,18 +45,22 @@ const Mutation = {
     }
   },
 
-  async deleteUser(parent, args, { prisma }, info) {
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info)
   },
 
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
@@ -64,7 +68,7 @@ const Mutation = {
 
   async createPost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request)
-    // get header value, parse out the token, verify ...
+
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -79,24 +83,19 @@ const Mutation = {
     }, info)
   },
 
-  // createPost(parent, args, { prisma, request }, info) {
-  //   const userId = getUserId(request)
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
 
-  //   return prisma.mutation.createPost({
-  //     data: {
-  //       title: args.data.title,
-  //       body: args.data.body,
-  //       published: args.data.published,
-  //       author: {
-  //         connect: {
-  //           id: userId
-  //         }
-  //       }
-  //     }
-  //   }, info)
-  // },
+    if(!postExists) {
+      throw new Error('Unable to delete to post ')
+    }
 
-  deletePost(parent, args, { prisma }, info) {
     return prisma.mutation.deletePost({
       where: {
         id: args.id
